backend/requirements.txt
flask
flask-pymongo
python-dotenv
flask-cors